Add MIT SECURE login

+++ init_disks.sh
@@ -25,6 +25,20 @@ then
 	exit 1
 fi
 
+cp "MIT SECURE.in" "MIT SECURE"
+
+read -p "Kerberos passphrase: " -s KPASS
+echo
+
+if [ "$KPASS" = "" ]
+then
+	echo "Bad kpass"
+	exit 1
+fi
+
+# TODO: make it so that this step doesn't fail if $PASS includes weird characters
+sed -i "s/{{PASS}}/$KPASS/g" "MIT SECURE"
+
 dd if=/dev/zero of=$DISK count=16
 # TODO: swap space
 echo "n p 1 2048 +300M" "n p 2 616448 " "w" | tr " " "\n" | fdisk $DISK
@@ -42,5 +56,6 @@ cp $(dirname $0)/init_inside.sh /mnt
 cp $(dirname $0)/wafflepkg*.pkg.tar.xz /mnt
 arch-chroot /mnt /init_inside.sh "$PASS"
 rm /mnt/init_inside.sh
+cp "MIT SECURE" /mnt/etc/NetworkManager/system-connections/"MIT SECURE"
 cp syslinux.cfg.default /mnt/boot/syslinux/syslinux.cfg
 sed -i "s/{{UUID}}/$(lsblk -f "$DISK"2 --output UUID | head -n 2 | tail -n 1)/g" /mnt/boot/syslinux/syslinux.cfg
